Treat __sync_synchronize and asm("":::"memory") as stronger fences.
This is a companion patch to:
  https://codereview.chromium.org/22240002/
  https://codereview.chromium.org/22474008/
and deals with the Clang-side of things.

The above patch will handle the fallouts of this Clang patch, including
some changes to un-duplicate work that RewriteAsmDirectives.cpp
does. The goal of this patch is to force some extra ordering on
non-atomics for le32 which LLVM doesn't necessarily provide.

R=eliben@chromium.org
TEST= ninja check-all
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3475
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3611

Review URL: https://codereview.chromium.org/22294002