Drop addAsmMemoryAroundSyncSynchronize() and asmMemoryIsFence().
For us, staying in sync with upstream on issues related to the memory model
is more important. None of clang, gcc, nor icc appear to offer these features.

If such features are desirable, they should ideally be implemented in upstream
clang, in a target-independent way, since this issue would affect users of all
platforms which support threads, not just PNaCl or Emscripten.