Driver: Fix handling of -stdlib= flag and link with pthread if using libc++
We don't plan to support libstdc++, so do not allow it to be selected via
the -stdlib= flag. (And in any case the current code makes it work for library flags
but not for headers). Also allows -stdlib=libc++ without emitting an
unused-argument warning.

Also our build of libc++ currently requires pthreads, so add -lpthread to
the CXX stdlib args.

R=noelallen@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3946

Review URL: https://codereview.chromium.org/636673003