Derive NaCl ToolChain from Generic_ELF instead of Linux
This prevents the GCC installation detector from injecting unwanted paths
(because it never runs), simplifies the NaCl-specific code, and fixes the
problem of host-dependent linker behavior which was caused by the Linux
constructor injecting different ExtraOps based on the distro.

Also updates the test to check for the --build-id flag and to check
the C++ header path.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3975

Review URL: https://codereview.chromium.org/775463002