Inject the NaCl ARM assembly macro file into ARM assembler jobs
ARM assembly can be written with a set of helper macros for the various
SFI requirements like register masking. Because it can be used by inline
assembly it needs to always be available. So inject it into all assembly
jobs for the ARM toolchain. This also matches nacl-gcc's behavior.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4018

Review URL: https://codereview.chromium.org/800393005