Refactor PNaCl target in Clang
A bug fix:
 - Atomic width was missing which caused it to be zero, leading Clang to try to emit library calls instead.

A few refactorings:
 - `this->` everywhere? Gone.
 - Some of the initializations weren't in the canonical order.
 - We were missing a description string.
 - The macro definitions weren't quite standard.

I also updated the test. The PNaCl target is now similar to other targets, especially the le64 target. Note that the test was made hermetic upstream, so I'm leaving it as-is for now (no need to cherrypick that part).

I can upstream this patch once we agree on it, and we could do the same for the NaCl targets.

See the related test patch:
  https://codereview.chromium.org/853823002/

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3941
TEST= ninja check-all
TEST= ./scons run_synchronization_sync_test bitcode=1
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/850763009