Inject gnueabihf environment into the effective clang triple for NaCl
The effective clang triple is what the clang driver passes as the
-target flag for its cc1 jobs. Setting gnueabihf as the default environment
for nacl means that the the triple with that environment will end up as the
module's triple (instead of just "armv7-nacl") and the clang LLVM sides will
always agree on the defaults. This mostly doesn't matter except for things
like calls to runtime functions such as those in compiler-rt or libcxxabi,
where a mismatch was observed.

With this change the modules have the environment in the module's triple
and the calling convention is not explicitly overriden on every function
define and callsite as it was before.

R=jfb@chromium.org, jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4018

Review URL: https://codereview.chromium.org/838933004