Use seperate usr/lib and usr/include directories for i686-nacl
This is what the naclports build system is expecting,
and I think its preferable anyway (for example, its
not a good idea to share include paths between
architectures in general).

In the long run I'd like to see lib32 move too, but
that is a bigger change.  This is just enough to make
naclports happy.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4108
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/946063002