[MIPS] Add support for direct-to-nacl in Clang
Cherry-picked r241678 from upstream, modified it to the current
PNaCl-Clang tree, and added an extra flag "-direct-to-nacl" that is not
present upstream.

Original commit message:

For Mips direct-to-nacl, the goal is to be close to le32 front-end and
use Mips32EL backend. This patch defines new NaClMips32ELTargetInfo and
modifies it slightly to be close to le32. It also adds necessary parts,
inline with ARM and X86.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4205
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/1208073002 .