Support use of external clang sources in emscripten-version number generation
Using clang sources external to LLVM (i.e. not in the tools/clang subdir)
breaks the includes used by fastcomp clang to print the version numbers.
This patch fixes the paths, but we also want for the time being not to break
any existing build scripts (such as emslave). So instead of just changing
the path, inject a define into the build to conditionally select one.

I didn't try to make it support separate versions for LLVM and clang
(since we dont actually use that currently) but I did keep the same
format.